Move DictionarySettings Tests from integration to unit
