+++ src/test/java/com/meilisearch/integration/SettingsTest.java
@@ -1057,54 +1057,6 @@ public class SettingsTest extends AbstractIT {
         assertThat(facetingAfterReset.getMaxValuesPerFacet(), is(equalTo(100)));
     }
 
-    /** Tests of the dictionary setting methods */
-    @Test
-    @DisplayName("Test get dictionary settings by uid")
-    public void testGetDictionarySettings() throws Exception {
-        Index index = createIndex("testGetDictionarySettings");
-        Settings initialSettings = index.getSettings();
-        String[] initialDictionary = index.getDictionarySettings();
-
-        assertThat(initialDictionary, is(arrayWithSize(initialSettings.getDictionary().length)));
-        assertThat(initialDictionary, is(equalTo(initialSettings.getDictionary())));
-    }
-
-    @Test
-    @DisplayName("Test update dictionary settings")
-    public void testUpdateDictionarySettings() throws Exception {
-        Index index = createIndex("testUpdateDictionarySettings");
-        String[] initialDictionary = index.getDictionarySettings();
-        String[] newDictionary = {"J. R. R.", "W. E. B."};
-
-        index.waitForTask(index.updateDictionarySettings(newDictionary).getTaskUid());
-        String[] updatedDictionarySettings = index.getDictionarySettings();
-
-        assertThat(updatedDictionarySettings, is(arrayWithSize(newDictionary.length)));
-        assertThat(updatedDictionarySettings, is(equalTo(newDictionary)));
-        assertThat(updatedDictionarySettings, is(not(arrayWithSize(initialDictionary.length))));
-    }
-
-    @Test
-    @DisplayName("Test reset dictionary settings")
-    public void testResetDictionarySettings() throws Exception {
-        Index index = createIndex("testResetDictionarySettings");
-        String[] initialDictionary = index.getDictionarySettings();
-        String[] newDictionary = {"J. R. R.", "W. E. B."};
-
-        index.waitForTask(index.updateDictionarySettings(newDictionary).getTaskUid());
-        String[] updatedDictionarySettings = index.getDictionarySettings();
-
-        index.waitForTask(index.resetDictionarySettings().getTaskUid());
-        String[] dictionaryAfterReset = index.getDictionarySettings();
-
-        assertThat(updatedDictionarySettings, is(arrayWithSize(newDictionary.length)));
-        assertThat(updatedDictionarySettings, is(equalTo(newDictionary)));
-        assertThat(updatedDictionarySettings, is(not(arrayWithSize(initialDictionary.length))));
-        assertThat(dictionaryAfterReset, is(not(arrayWithSize(updatedDictionarySettings.length))));
-        assertThat(dictionaryAfterReset, is(arrayWithSize(initialDictionary.length)));
-        assertThat(dictionaryAfterReset, is(equalTo(initialDictionary)));
-    }
-
     private Index createIndex(String indexUid) throws Exception {
         Index index = client.index(indexUid);
         TaskInfo updateInfo = index.addDocuments(testData.getRaw());
